Regulator/requirements.txt
pytest
pytest-timeout
paho-mqtt
kafka-python
pydantic
python-dotenv